pyproject.toml
[project]
name = "Moodle_RAG_with_Crowler"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromium>=0.0.0",
    "crawl4ai>=0.8.0",
    "deep-translator>=1.11.4",
    "python-dotenv>=1.2.1",
    "langchain-chroma>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-text-splitters>=1.1.0",
    "langdetect>=1.0.9",
    "mlx-lm>=0.30.7",
    "pandas>=3.0.1",
    "pip>=26.0.1",
    "playwright>=1.58.0",
    "sentence-transformers>=5.2.3",
    "torch>=2.10.0",
    "fastapi>=0.129.0",
    "uvicorn>=0.41.0",
]

[dependency-groups]
dev = [
    "mypy>=1.19.1",
    "ruff>=0.15.2",
]

[tool.ruff]
line-length = 120
target-version = "py312"
src = ["scripts"]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (ruff сам форматирует)
    "B008",  # do not perform function calls in argument defaults
]

[tool.ruff.lint.isort]
known-first-party = ["scripts"]

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_configs = true

[[tool.mypy.overrides]]
module = [
    "crawl4ai.*",
    "langchain_chroma.*",
    "langchain_community.*",
    "langchain_text_splitters.*",
    "langdetect.*",
    "deep_translator.*",
    "mlx_lm.*",
]
ignore_missing_imports = true
